Split vec and set into multiple files
