+++ set/RELEASES.md
@@ -1,4 +1,25 @@
-Version 0.7.0 (not yet released) (ZERO BREAKING CHANGES)
+Version 0.10.0 (not yet released)
+========================================================
+
+<a id="v0.10.0"></a>
+
+Version 0.9.0
+========================================================
+
+<a id="v0.9.0"></a>
+
+- Minimal Supported Rust Version is 1.82
+- Rust edition 2021 is used
+- implemented `fn make_empty`
+- implemented `fn reset`
+- added general initialization functions: `fn new_general`, `fn from_bit_vec_general`, `fn with_capacity_general`, `fn from_bytes_general`
+
+Version 0.8.0
+========================================================
+
+<a id="v0.8.0"></a>
+
+Version 0.7.0 (ZERO BREAKING CHANGES)
 ========================================================
 
 <a id="v0.7.0"></a>
