+++ set/src/lib.rs
@@ -57,1166 +57,22 @@
 #[cfg(any(test, feature = "std"))]
 extern crate std;
 
-use bit_vec::{BitBlock, BitVec, Blocks};
-use bit_vec::{BitBlockOrStore, BitStore};
-use core::cmp;
-use core::cmp::Ordering;
-use core::fmt;
-use core::hash;
-use core::iter::{self, Chain, Enumerate, FromIterator, Repeat, Skip, Take};
-
 #[cfg(feature = "nanoserde")]
 extern crate alloc;
-#[cfg(feature = "nanoserde")]
-use alloc::vec::Vec;
-#[cfg(feature = "nanoserde")]
-use nanoserde::{DeBin, DeJson, DeRon, SerBin, SerJson, SerRon};
 
-#[allow(type_alias_bounds)]
-type Block<B: BitBlockOrStore> = <B::Store as BitStore>::Block;
-#[allow(type_alias_bounds)]
-type MatchWords<'a, B: BitBlockOrStore> =
-    Chain<Enumerate<Blocks<'a, B>>, Skip<Take<Enumerate<Repeat<Block<B>>>>>>;
+mod set;
+mod util;
+mod iter;
 
-/// Computes how many blocks are needed to store that many bits
-fn blocks_for_bits<B: BitBlockOrStore>(bits: usize) -> usize {
-    // If we want 17 bits, dividing by 32 will produce 0. So we add 1 to make sure we
-    // reserve enough. But if we want exactly a multiple of 32, this will actually allocate
-    // one too many. So we need to check if that's the case. We can do that by computing if
-    // bitwise AND by `32 - 1` is 0. But LLVM should be able to optimize the semantically
-    // superior modulo operator on a power of two to this.
-    //
-    // Note that we can technically avoid this branch with the expression
-    // `(nbits + BITS - 1) / 32::BITS`, but if nbits is almost usize::MAX this will overflow.
-    if bits % B::BITS == 0 {
-        bits / B::BITS
-    } else {
-        bits / B::BITS + 1
-    }
+pub mod local_prelude {
+    pub use bit_vec::{BitBlock, BitBlockOrStore, BitStore, BitVec, Blocks};
+    pub use core::cmp::Ordering;
+    pub use core::{hash, fmt, cmp};
+    pub use core::iter::{self, Chain, Enumerate, FromIterator, Repeat, Skip, Take};
 }
 
-#[allow(clippy::iter_skip_zero)]
-// Take two BitVec's, and return iterators of their words, where the shorter one
-// has been padded with 0's
-fn match_words<'a, 'b, B: BitBlockOrStore>(
-    a: &'a BitVec<B>,
-    b: &'b BitVec<B>,
-) -> (MatchWords<'a, B>, MatchWords<'b, B>) {
-    let a_len = a.storage().len();
-    let b_len = b.storage().len();
-
-    // have to uselessly pretend to pad the longer one for type matching
-    if a_len < b_len {
-        (
-            a.blocks()
-                .enumerate()
-                .chain(iter::repeat(B::ZERO).enumerate().take(b_len).skip(a_len)),
-            b.blocks()
-                .enumerate()
-                .chain(iter::repeat(B::ZERO).enumerate().take(0).skip(0)),
-        )
-    } else {
-        (
-            a.blocks()
-                .enumerate()
-                .chain(iter::repeat(B::ZERO).enumerate().take(0).skip(0)),
-            b.blocks()
-                .enumerate()
-                .chain(iter::repeat(B::ZERO).enumerate().take(a_len).skip(b_len)),
-        )
-    }
-}
-
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-#[cfg_attr(
-    feature = "borsh",
-    derive(borsh::BorshDeserialize, borsh::BorshSerialize)
-)]
-#[cfg_attr(
-    feature = "miniserde",
-    derive(miniserde::Deserialize, miniserde::Serialize)
-)]
-#[cfg_attr(
-    feature = "nanoserde",
-    derive(DeBin, DeJson, DeRon, SerBin, SerJson, SerRon)
-)]
-pub struct BitSet<B: BitBlockOrStore = u32> {
-    bit_vec: BitVec<B>,
-}
-
-impl<B: BitBlockOrStore> Clone for BitSet<B> {
-    fn clone(&self) -> Self {
-        BitSet {
-            bit_vec: self.bit_vec.clone(),
-        }
-    }
-
-    fn clone_from(&mut self, other: &Self) {
-        self.bit_vec.clone_from(&other.bit_vec);
-    }
-}
-
-impl<B: BitBlockOrStore> Default for BitSet<B> {
-    #[inline]
-    fn default() -> Self {
-        BitSet {
-            bit_vec: Default::default(),
-        }
-    }
-}
-
-impl<B: BitBlockOrStore> FromIterator<usize> for BitSet<B> {
-    fn from_iter<I: IntoIterator<Item = usize>>(iter: I) -> Self {
-        let mut ret = Self::default();
-        ret.extend(iter);
-        ret
-    }
-}
-
-impl<B: BitBlockOrStore> Extend<usize> for BitSet<B> {
-    #[inline]
-    fn extend<I: IntoIterator<Item = usize>>(&mut self, iter: I) {
-        for i in iter {
-            self.insert(i);
-        }
-    }
-}
-
-impl<B: BitBlockOrStore> PartialOrd for BitSet<B> {
-    #[inline]
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl<B: BitBlockOrStore> Ord for BitSet<B> {
-    #[inline]
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.iter().cmp(other)
-    }
-}
-
-impl<B: BitBlockOrStore> PartialEq for BitSet<B> {
-    #[inline]
-    fn eq(&self, other: &Self) -> bool {
-        self.iter().eq(other)
-    }
-}
-
-impl<B: BitBlockOrStore> Eq for BitSet<B> {}
-
-impl BitSet<u32> {
-    /// Creates a new empty `BitSet`.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let mut s = BitSet::new();
-    /// ```
-    #[inline]
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Creates a new `BitSet` with initially no contents, able to
-    /// hold `nbits` elements without resizing.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let mut s = BitSet::with_capacity(100);
-    /// assert!(s.capacity() >= 100);
-    /// ```
-    #[inline]
-    pub fn with_capacity(nbits: usize) -> Self {
-        let bit_vec = BitVec::from_elem(nbits, false);
-        Self::from_bit_vec(bit_vec)
-    }
-
-    /// Creates a new `BitSet` from the given bit vector.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_vec::BitVec;
-    /// use bit_set::BitSet;
-    ///
-    /// let bv = BitVec::from_bytes(&[0b01100000]);
-    /// let s = BitSet::from_bit_vec(bv);
-    ///
-    /// // Print 1, 2 in arbitrary order
-    /// for x in s.iter() {
-    ///     println!("{}", x);
-    /// }
-    /// ```
-    #[inline]
-    pub fn from_bit_vec(bit_vec: BitVec) -> Self {
-        BitSet { bit_vec }
-    }
-
-    pub fn from_bytes(bytes: &[u8]) -> Self {
-        BitSet {
-            bit_vec: BitVec::from_bytes(bytes),
-        }
-    }
-}
-
-impl<B: BitBlockOrStore> BitSet<B> {
-    /// Creates a new empty `BitSet`.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let mut s = <BitSet>::new_general();
-    /// ```
-    #[inline]
-    pub fn new_general() -> Self {
-        Self::default()
-    }
-
-    /// Creates a new `BitSet` with initially no contents, able to
-    /// hold `nbits` elements without resizing.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let mut s = <BitSet>::with_capacity_general(100);
-    /// assert!(s.capacity() >= 100);
-    /// ```
-    #[inline]
-    pub fn with_capacity_general(nbits: usize) -> Self {
-        let bit_vec = BitVec::from_elem_general(nbits, false);
-        Self::from_bit_vec_general(bit_vec)
-    }
-
-    /// Creates a new `BitSet` from the given bit vector.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_vec::BitVec;
-    /// use bit_set::BitSet;
-    ///
-    /// let bv: BitVec<u64> = BitVec::from_bytes_general(&[0b01100000]);
-    /// let s = BitSet::from_bit_vec_general(bv);
-    ///
-    /// // Print 1, 2 in arbitrary order
-    /// for x in s.iter() {
-    ///     println!("{}", x);
-    /// }
-    /// ```
-    #[inline]
-    pub fn from_bit_vec_general(bit_vec: BitVec<B>) -> Self {
-        BitSet { bit_vec }
-    }
-
-    pub fn from_bytes_general(bytes: &[u8]) -> Self {
-        BitSet {
-            bit_vec: BitVec::from_bytes_general(bytes),
-        }
-    }
-
-    /// Returns the capacity in bits for this bit vector. Inserting any
-    /// element less than this amount will not trigger a resizing.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let mut s = BitSet::with_capacity(100);
-    /// assert!(s.capacity() >= 100);
-    /// ```
-    #[inline]
-    pub fn capacity(&self) -> usize {
-        self.bit_vec.capacity()
-    }
-
-    /// Reserves capacity for the given `BitSet` to contain `len` distinct elements. In the case
-    /// of `BitSet` this means reallocations will not occur as long as all inserted elements
-    /// are less than `len`.
-    ///
-    /// The collection may reserve more space to avoid frequent reallocations.
-    ///
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let mut s = BitSet::new();
-    /// s.reserve_len(10);
-    /// assert!(s.capacity() >= 10);
-    /// ```
-    pub fn reserve_len(&mut self, len: usize) {
-        let cur_len = self.bit_vec.len();
-        if len >= cur_len {
-            self.bit_vec.reserve(len - cur_len);
-        }
-    }
-
-    /// Reserves the minimum capacity for the given `BitSet` to contain `len` distinct elements.
-    /// In the case of `BitSet` this means reallocations will not occur as long as all inserted
-    /// elements are less than `len`.
-    ///
-    /// Note that the allocator may give the collection more space than it requests. Therefore
-    /// capacity can not be relied upon to be precisely minimal. Prefer `reserve_len` if future
-    /// insertions are expected.
-    ///
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let mut s = BitSet::new();
-    /// s.reserve_len_exact(10);
-    /// assert!(s.capacity() >= 10);
-    /// ```
-    pub fn reserve_len_exact(&mut self, len: usize) {
-        let cur_len = self.bit_vec.len();
-        if len >= cur_len {
-            self.bit_vec.reserve_exact(len - cur_len);
-        }
-    }
-
-    /// Consumes this set to return the underlying bit vector.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let mut s = BitSet::new();
-    /// s.insert(0);
-    /// s.insert(3);
-    ///
-    /// let bv = s.into_bit_vec();
-    /// assert!(bv[0]);
-    /// assert!(bv[3]);
-    /// ```
-    #[inline]
-    pub fn into_bit_vec(self) -> BitVec<B> {
-        self.bit_vec
-    }
-
-    /// Returns a reference to the underlying bit vector.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let mut set = BitSet::new();
-    /// set.insert(0);
-    ///
-    /// let bv = set.get_ref();
-    /// assert_eq!(bv[0], true);
-    /// ```
-    #[inline]
-    pub fn get_ref(&self) -> &BitVec<B> {
-        &self.bit_vec
-    }
-
-    /// Returns a mutable reference to the underlying bit vector.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let mut set = BitSet::new();
-    /// set.insert(0);
-    /// set.insert(3);
-    ///
-    /// {
-    ///     let bv = set.get_mut();
-    ///     bv.set(1, true);
-    /// }
-    ///
-    /// assert!(set.contains(0));
-    /// assert!(set.contains(1));
-    /// assert!(set.contains(3));
-    /// ```
-    #[inline]
-    pub fn get_mut(&mut self) -> &mut BitVec<B> {
-        &mut self.bit_vec
-    }
-
-    #[inline]
-    fn other_op<F>(&mut self, other: &Self, mut f: F)
-    where
-        F: FnMut(Block<B>, Block<B>) -> Block<B>,
-    {
-        // Unwrap BitVecs
-        let self_bit_vec = &mut self.bit_vec;
-        let other_bit_vec = &other.bit_vec;
-
-        let self_len = self_bit_vec.len();
-        let other_len = other_bit_vec.len();
-
-        // Expand the vector if necessary
-        if self_len < other_len {
-            self_bit_vec.grow(other_len - self_len, false);
-        }
-
-        // virtually pad other with 0's for equal lengths
-        let other_words = {
-            let (_, result) = match_words(self_bit_vec, other_bit_vec);
-            result
-        };
-
-        // Apply values found in other
-        for (i, w) in other_words {
-            let old = self_bit_vec.storage()[i];
-            let new = f(old, w);
-            unsafe {
-                self_bit_vec.storage_mut().slice_mut()[i] = new;
-            }
-        }
-    }
-
-    /// Truncates the underlying vector to the least length required.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let mut s = BitSet::new();
-    /// s.insert(3231);
-    /// s.remove(3231);
-    ///
-    /// // Internal storage will probably be bigger than necessary
-    /// println!("old capacity: {}", s.capacity());
-    /// assert!(s.capacity() >= 3231);
-    ///
-    /// // Now should be smaller
-    /// s.shrink_to_fit();
-    /// println!("new capacity: {}", s.capacity());
-    /// ```
-    #[inline]
-    pub fn shrink_to_fit(&mut self) {
-        let bit_vec = &mut self.bit_vec;
-        // Obtain original length
-        let old_len = bit_vec.storage().len();
-        // Obtain coarse trailing zero length
-        let n = bit_vec
-            .storage()
-            .iter()
-            .rev()
-            .take_while(|&&n| n == B::ZERO)
-            .count();
-        // Truncate away all empty trailing blocks, then shrink_to_fit
-        let trunc_len = old_len - n;
-        unsafe {
-            bit_vec.storage_mut().truncate(trunc_len);
-            bit_vec.set_len(trunc_len * B::BITS);
-        }
-        bit_vec.shrink_to_fit();
-    }
-
-    /// Iterator over each usize stored in the `BitSet`.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let s = BitSet::from_bytes(&[0b01001010]);
-    ///
-    /// // Print 1, 4, 6 in arbitrary order
-    /// for x in s.iter() {
-    ///     println!("{}", x);
-    /// }
-    /// ```
-    #[inline]
-    pub fn iter(&self) -> Iter<'_, B> {
-        Iter(BlockIter::from_blocks(self.bit_vec.blocks()))
-    }
-
-    /// Iterator over each usize stored in `self` union `other`.
-    /// See [`union_with`] for an efficient in-place version.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let a = BitSet::from_bytes(&[0b01101000]);
-    /// let b = BitSet::from_bytes(&[0b10100000]);
-    ///
-    /// // Print 0, 1, 2, 4 in arbitrary order
-    /// for x in a.union(&b) {
-    ///     println!("{}", x);
-    /// }
-    /// ```
-    ///
-    /// [`union_with`]: Self::union_with
-    #[inline]
-    pub fn union<'a>(&'a self, other: &'a Self) -> Union<'a, B> {
-        fn or<B: BitBlock>(w1: B, w2: B) -> B {
-            w1 | w2
-        }
-
-        Union(BlockIter::from_blocks(TwoBitPositions {
-            set: self.bit_vec.blocks(),
-            other: other.bit_vec.blocks(),
-            merge: or,
-        }))
-    }
-
-    /// Iterator over each usize stored in `self` intersect `other`.
-    /// See [`intersect_with`] for an efficient in-place version.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let a = BitSet::from_bytes(&[0b01101000]);
-    /// let b = BitSet::from_bytes(&[0b10100000]);
-    ///
-    /// // Print 2
-    /// for x in a.intersection(&b) {
-    ///     println!("{}", x);
-    /// }
-    /// ```
-    ///
-    /// [`intersect_with`]: Self::intersect_with
-    #[inline]
-    pub fn intersection<'a>(&'a self, other: &'a Self) -> Intersection<'a, B> {
-        fn bitand<B: BitBlock>(w1: B, w2: B) -> B {
-            w1 & w2
-        }
-        let min = cmp::min(self.bit_vec.len(), other.bit_vec.len());
-
-        Intersection {
-            iter: BlockIter::from_blocks(TwoBitPositions {
-                set: self.bit_vec.blocks(),
-                other: other.bit_vec.blocks(),
-                merge: bitand,
-            }),
-            n: min,
-        }
-    }
-
-    /// Iterator over each usize stored in the `self` setminus `other`.
-    /// See [`difference_with`] for an efficient in-place version.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let a = BitSet::from_bytes(&[0b01101000]);
-    /// let b = BitSet::from_bytes(&[0b10100000]);
-    ///
-    /// // Print 1, 4 in arbitrary order
-    /// for x in a.difference(&b) {
-    ///     println!("{}", x);
-    /// }
-    ///
-    /// // Note that difference is not symmetric,
-    /// // and `b - a` means something else.
-    /// // This prints 0
-    /// for x in b.difference(&a) {
-    ///     println!("{}", x);
-    /// }
-    /// ```
-    ///
-    /// [`difference_with`]: Self::difference_with
-    #[inline]
-    pub fn difference<'a>(&'a self, other: &'a Self) -> Difference<'a, B> {
-        fn diff<B: BitBlock>(w1: B, w2: B) -> B {
-            w1 & !w2
-        }
-
-        Difference(BlockIter::from_blocks(TwoBitPositions {
-            set: self.bit_vec.blocks(),
-            other: other.bit_vec.blocks(),
-            merge: diff,
-        }))
-    }
-
-    /// Iterator over each usize stored in the symmetric difference of `self` and `other`.
-    /// See [`symmetric_difference_with`] for an efficient in-place version.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let a = BitSet::from_bytes(&[0b01101000]);
-    /// let b = BitSet::from_bytes(&[0b10100000]);
-    ///
-    /// // Print 0, 1, 4 in arbitrary order
-    /// for x in a.symmetric_difference(&b) {
-    ///     println!("{}", x);
-    /// }
-    /// ```
-    ///
-    /// [`symmetric_difference_with`]: Self::symmetric_difference_with
-    #[inline]
-    pub fn symmetric_difference<'a>(&'a self, other: &'a Self) -> SymmetricDifference<'a, B> {
-        fn bitxor<B: BitBlock>(w1: B, w2: B) -> B {
-            w1 ^ w2
-        }
-
-        SymmetricDifference(BlockIter::from_blocks(TwoBitPositions {
-            set: self.bit_vec.blocks(),
-            other: other.bit_vec.blocks(),
-            merge: bitxor,
-        }))
-    }
-
-    /// Unions in-place with the specified other bit vector.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let a   = 0b01101000;
-    /// let b   = 0b10100000;
-    /// let res = 0b11101000;
-    ///
-    /// let mut a = BitSet::from_bytes(&[a]);
-    /// let b = BitSet::from_bytes(&[b]);
-    /// let res = BitSet::from_bytes(&[res]);
-    ///
-    /// a.union_with(&b);
-    /// assert_eq!(a, res);
-    /// ```
-    #[inline]
-    pub fn union_with(&mut self, other: &Self) {
-        self.other_op(other, |w1, w2| w1 | w2);
-    }
-
-    /// Intersects in-place with the specified other bit vector.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let a   = 0b01101000;
-    /// let b   = 0b10100000;
-    /// let res = 0b00100000;
-    ///
-    /// let mut a = BitSet::from_bytes(&[a]);
-    /// let b = BitSet::from_bytes(&[b]);
-    /// let res = BitSet::from_bytes(&[res]);
-    ///
-    /// a.intersect_with(&b);
-    /// assert_eq!(a, res);
-    /// ```
-    #[inline]
-    pub fn intersect_with(&mut self, other: &Self) {
-        self.other_op(other, |w1, w2| w1 & w2);
-    }
-
-    /// Makes this bit vector the difference with the specified other bit vector
-    /// in-place.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let a   = 0b01101000;
-    /// let b   = 0b10100000;
-    /// let a_b = 0b01001000; // a - b
-    /// let b_a = 0b10000000; // b - a
-    ///
-    /// let mut bva = BitSet::from_bytes(&[a]);
-    /// let bvb = BitSet::from_bytes(&[b]);
-    /// let bva_b = BitSet::from_bytes(&[a_b]);
-    /// let bvb_a = BitSet::from_bytes(&[b_a]);
-    ///
-    /// bva.difference_with(&bvb);
-    /// assert_eq!(bva, bva_b);
-    ///
-    /// let bva = BitSet::from_bytes(&[a]);
-    /// let mut bvb = BitSet::from_bytes(&[b]);
-    ///
-    /// bvb.difference_with(&bva);
-    /// assert_eq!(bvb, bvb_a);
-    /// ```
-    #[inline]
-    pub fn difference_with(&mut self, other: &Self) {
-        self.other_op(other, |w1, w2| w1 & !w2);
-    }
-
-    /// Makes this bit vector the symmetric difference with the specified other
-    /// bit vector in-place.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bit_set::BitSet;
-    ///
-    /// let a   = 0b01101000;
-    /// let b   = 0b10100000;
-    /// let res = 0b11001000;
-    ///
-    /// let mut a = BitSet::from_bytes(&[a]);
-    /// let b = BitSet::from_bytes(&[b]);
-    /// let res = BitSet::from_bytes(&[res]);
-    ///
-    /// a.symmetric_difference_with(&b);
-    /// assert_eq!(a, res);
-    /// ```
-    #[inline]
-    pub fn symmetric_difference_with(&mut self, other: &Self) {
-        self.other_op(other, |w1, w2| w1 ^ w2);
-    }
-
-    /*
-        /// Moves all elements from `other` into `Self`, leaving `other` empty.
-        ///
-        /// # Examples
-        ///
-        /// ```
-        /// use bit_set::BitSet;
-        ///
-        /// let mut a = BitSet::new();
-        /// a.insert(2);
-        /// a.insert(6);
-        ///
-        /// let mut b = BitSet::new();
-        /// b.insert(1);
-        /// b.insert(3);
-        /// b.insert(6);
-        ///
-        /// a.append(&mut b);
-        ///
-        /// assert_eq!(a.len(), 4);
-        /// assert_eq!(b.len(), 0);
-        /// assert_eq!(a, BitSet::from_bytes(&[0b01110010]));
-        /// ```
-        pub fn append(&mut self, other: &mut Self) {
-            self.union_with(other);
-            other.clear();
-        }
-
-        /// Splits the `BitSet` into two at the given key including the key.
-        /// Retains the first part in-place while returning the second part.
-        ///
-        /// # Examples
-        ///
-        /// ```
-        /// use bit_set::BitSet;
-        ///
-        /// let mut a = BitSet::new();
-        /// a.insert(2);
-        /// a.insert(6);
-        /// a.insert(1);
-        /// a.insert(3);
-        ///
-        /// let b = a.split_off(3);
-        ///
-        /// assert_eq!(a.len(), 2);
-        /// assert_eq!(b.len(), 2);
-        /// assert_eq!(a, BitSet::from_bytes(&[0b01100000]));
-        /// assert_eq!(b, BitSet::from_bytes(&[0b00010010]));
-        /// ```
-        pub fn split_off(&mut self, at: usize) -> Self {
-            let mut other = BitSet::new();
-
-            if at == 0 {
-                swap(self, &mut other);
-                return other;
-            } else if at >= self.bit_vec.len() {
-                return other;
-            }
-
-            // Calculate block and bit at which to split
-            let w = at / BITS;
-            let b = at % BITS;
-
-            // Pad `other` with `w` zero blocks,
-            // append `self`'s blocks in the range from `w` to the end to `other`
-            other.bit_vec.storage_mut().extend(repeat(0u32).take(w)
-                                         .chain(self.bit_vec.storage()[w..].iter().cloned()));
-            other.bit_vec.nbits = self.bit_vec.nbits;
-
-            if b > 0 {
-                other.bit_vec.storage_mut()[w] &= !0 << b;
-            }
-
-            // Sets `bit_vec.len()` and fixes the last block as well
-            self.bit_vec.truncate(at);
-
-            other
-        }
-    */
-
-    /// Counts the number of set bits in this set.
-    ///
-    /// Note that this function scans the set to calculate the number.
-    #[inline]
-    pub fn count(&self) -> usize {
-        self.bit_vec.blocks().fold(0, |acc, n| acc + n.count_ones())
-    }
-
-    /// Counts the number of set bits in this set.
-    ///
-    /// Note that this function scans the set to calculate the number.
-    #[inline]
-    #[deprecated = "use BitVec::count() instead"]
-    pub fn len(&self) -> usize {
-        self.count()
-    }
-
-    /// Returns whether there are no bits set in this set
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.bit_vec.none()
-    }
-
-    /// Removes all elements of this set.
-    ///
-    /// Different from [`reset`] only in that the capacity is preserved.
-    ///
-    /// [`reset`]: Self::reset
-    #[inline]
-    pub fn make_empty(&mut self) {
-        self.bit_vec.fill(false);
-    }
-
-    /// Resets this set to an empty state.
-    ///
-    /// Different from [`make_empty`] only in that the capacity may NOT be preserved.
-    ///
-    /// [`make_empty`]: Self::make_empty
-    #[inline]
-    pub fn reset(&mut self) {
-        self.bit_vec.remove_all();
-    }
-
-    /// Clears all bits in this set
-    #[deprecated(since = "0.9.0", note = "please use `fn make_empty` instead")]
-    #[inline]
-    pub fn clear(&mut self) {
-        self.make_empty();
-    }
-
-    /// Returns `true` if this set contains the specified integer.
-    #[inline]
-    pub fn contains(&self, value: usize) -> bool {
-        let bit_vec = &self.bit_vec;
-        value < bit_vec.len() && bit_vec[value]
-    }
-
-    /// Returns `true` if the set has no elements in common with `other`.
-    /// This is equivalent to checking for an empty intersection.
-    #[inline]
-    pub fn is_disjoint(&self, other: &Self) -> bool {
-        self.intersection(other).next().is_none()
-    }
-
-    /// Returns `true` if the set is a subset of another.
-    #[inline]
-    pub fn is_subset(&self, other: &Self) -> bool {
-        let self_bit_vec = &self.bit_vec;
-        let other_bit_vec = &other.bit_vec;
-        let other_blocks = blocks_for_bits::<B>(other_bit_vec.len());
-
-        // Check that `self` intersect `other` is self
-        self_bit_vec.blocks().zip(other_bit_vec.blocks()).all(|(w1, w2)| w1 & w2 == w1) &&
-        // Make sure if `self` has any more blocks than `other`, they're all 0
-        self_bit_vec.blocks().skip(other_blocks).all(|w| w == B::ZERO)
-    }
-
-    /// Returns `true` if the set is a superset of another.
-    #[inline]
-    pub fn is_superset(&self, other: &Self) -> bool {
-        other.is_subset(self)
-    }
-
-    /// Adds a value to the set. Returns `true` if the value was not already
-    /// present in the set.
-    pub fn insert(&mut self, value: usize) -> bool {
-        if self.contains(value) {
-            return false;
-        }
-
-        // Ensure we have enough space to hold the new element
-        let len = self.bit_vec.len();
-        if value >= len {
-            self.bit_vec.grow(value - len + 1, false);
-        }
-
-        self.bit_vec.set(value, true);
-        true
-    }
-
-    /// Removes a value from the set. Returns `true` if the value was
-    /// present in the set.
-    pub fn remove(&mut self, value: usize) -> bool {
-        if !self.contains(value) {
-            return false;
-        }
-
-        self.bit_vec.set(value, false);
-
-        true
-    }
-
-    /// Excludes `element` and all greater elements from the `BitSet`.
-    pub fn truncate(&mut self, element: usize) {
-        self.bit_vec.truncate(element);
-    }
-}
-
-impl<B: BitBlockOrStore> fmt::Debug for BitSet<B> {
-    fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
-        fmt.debug_struct("BitSet")
-            .field("bit_vec", &self.bit_vec)
-            .finish()
-    }
-}
-
-impl<B: BitBlockOrStore> fmt::Display for BitSet<B> {
-    fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
-        fmt.debug_set().entries(self).finish()
-    }
-}
-
-impl<B: BitBlockOrStore> hash::Hash for BitSet<B> {
-    fn hash<H: hash::Hasher>(&self, state: &mut H) {
-        for pos in self {
-            pos.hash(state);
-        }
-    }
-}
-
-#[derive(Clone)]
-struct BlockIter<T, B: BitBlockOrStore> {
-    head: Block<B>,
-    head_offset: usize,
-    tail: T,
-}
-
-impl<T, B: BitBlockOrStore> BlockIter<T, B>
-where
-    T: Iterator<Item = Block<B>>,
-{
-    fn from_blocks(mut blocks: T) -> Self {
-        let h = blocks.next().unwrap_or(B::ZERO);
-        BlockIter {
-            tail: blocks,
-            head: h,
-            head_offset: 0,
-        }
-    }
-}
-
-/// An iterator combining two `BitSet` iterators.
-#[derive(Clone)]
-struct TwoBitPositions<'a, B: 'a + BitBlockOrStore> {
-    set: Blocks<'a, B>,
-    other: Blocks<'a, B>,
-    merge: fn(Block<B>, Block<B>) -> Block<B>,
-}
-
-/// An iterator for `BitSet`.
-#[derive(Clone)]
-pub struct Iter<'a, B: 'a + BitBlockOrStore>(BlockIter<Blocks<'a, B>, B>);
-#[derive(Clone)]
-pub struct Union<'a, B: 'a + BitBlockOrStore>(BlockIter<TwoBitPositions<'a, B>, B>);
-#[derive(Clone)]
-pub struct Intersection<'a, B: 'a + BitBlockOrStore> {
-    iter: BlockIter<TwoBitPositions<'a, B>, B>,
-    // as an optimization, we compute the maximum possible
-    // number of elements in the intersection, and count it
-    // down as we return elements. If we reach zero, we can
-    // stop.
-    n: usize,
-}
-#[derive(Clone)]
-pub struct Difference<'a, B: 'a + BitBlockOrStore>(BlockIter<TwoBitPositions<'a, B>, B>);
-#[derive(Clone)]
-pub struct SymmetricDifference<'a, B: 'a + BitBlockOrStore>(BlockIter<TwoBitPositions<'a, B>, B>);
-
-impl<T, B: BitBlockOrStore> Iterator for BlockIter<T, B>
-where
-    T: Iterator<Item = Block<B>>,
-{
-    type Item = usize;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        while self.head == B::ZERO {
-            match self.tail.next() {
-                Some(w) => self.head = w,
-                None => return None,
-            }
-            self.head_offset += B::BITS;
-        }
-
-        // from the current block, isolate the
-        // LSB and subtract 1, producing k:
-        // a block with a number of set bits
-        // equal to the index of the LSB
-        let k = (self.head & (!self.head + B::ONE)) - B::ONE;
-        // update block, removing the LSB
-        self.head = self.head & (self.head - B::ONE);
-        // return offset + (index of LSB)
-        Some(self.head_offset + (<B::Store as BitStore>::Block::count_ones(k)))
-    }
-
-    fn count(self) -> usize {
-        self.head.count_ones() + self.tail.map(|block| block.count_ones()).sum::<usize>()
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        match self.tail.size_hint() {
-            (_, Some(h)) => (0, Some((1 + h) * B::BITS)),
-            _ => (0, None),
-        }
-    }
-}
-
-impl<B: BitBlockOrStore> Iterator for TwoBitPositions<'_, B> {
-    type Item = Block<B>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        match (self.set.next(), self.other.next()) {
-            (Some(a), Some(b)) => Some((self.merge)(a, b)),
-            (Some(a), None) => Some((self.merge)(a, B::ZERO)),
-            (None, Some(b)) => Some((self.merge)(B::ZERO, b)),
-            _ => None,
-        }
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        let (first_lower_bound, first_upper_bound) = self.set.size_hint();
-        let (second_lower_bound, second_upper_bound) = self.other.size_hint();
-
-        let upper_bound = first_upper_bound.zip(second_upper_bound);
-
-        let get_max = |(a, b)| cmp::max(a, b);
-        (
-            cmp::max(first_lower_bound, second_lower_bound),
-            upper_bound.map(get_max),
-        )
-    }
-}
-
-impl<B: BitBlockOrStore> Iterator for Iter<'_, B> {
-    type Item = usize;
-
-    #[inline]
-    fn next(&mut self) -> Option<usize> {
-        self.0.next()
-    }
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.0.size_hint()
-    }
-    #[inline]
-    fn count(self) -> usize {
-        self.0.count()
-    }
-}
-
-impl<B: BitBlockOrStore> Iterator for Union<'_, B> {
-    type Item = usize;
-
-    #[inline]
-    fn next(&mut self) -> Option<usize> {
-        self.0.next()
-    }
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.0.size_hint()
-    }
-    #[inline]
-    fn count(self) -> usize {
-        self.0.count()
-    }
-}
-
-impl<B: BitBlockOrStore> Iterator for Intersection<'_, B> {
-    type Item = usize;
-
-    #[inline]
-    fn next(&mut self) -> Option<usize> {
-        if self.n != 0 {
-            self.n -= 1;
-            self.iter.next()
-        } else {
-            None
-        }
-    }
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        // We could invoke self.iter.size_hint() and incorporate that into the hint.
-        // In practice, that does not seem worthwhile because the lower bound will
-        // always be zero and the upper bound could only possibly less then n in a
-        // partially iterated iterator. However, it makes little sense ask for size_hint
-        // in a partially iterated iterator, so it did not seem worthwhile.
-        (0, Some(self.n))
-    }
-    #[inline]
-    fn count(self) -> usize {
-        self.iter.count()
-    }
-}
-
-impl<B: BitBlockOrStore> Iterator for Difference<'_, B> {
-    type Item = usize;
-
-    #[inline]
-    fn next(&mut self) -> Option<usize> {
-        self.0.next()
-    }
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.0.size_hint()
-    }
-    #[inline]
-    fn count(self) -> usize {
-        self.0.count()
-    }
-}
-
-impl<B: BitBlockOrStore> Iterator for SymmetricDifference<'_, B> {
-    type Item = usize;
-
-    #[inline]
-    fn next(&mut self) -> Option<usize> {
-        self.0.next()
-    }
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.0.size_hint()
-    }
-    #[inline]
-    fn count(self) -> usize {
-        self.0.count()
-    }
-}
-
-impl<'a, B: BitBlockOrStore> IntoIterator for &'a BitSet<B> {
-    type Item = usize;
-    type IntoIter = Iter<'a, B>;
-
-    fn into_iter(self) -> Iter<'a, B> {
-        self.iter()
-    }
-}
+pub use set::BitSet;
+pub use bit_vec::{BitStore, BitBlockOrStore};
 
 #[cfg(test)]
 mod tests {
@@ -1224,7 +80,7 @@ mod tests {
     #![allow(clippy::shadow_same)]
     #![allow(clippy::shadow_unrelated)]
 
-    use super::BitSet;
+    use crate::set::BitSet;
     use bit_vec::BitVec;
     use std::cmp::Ordering::{Equal, Greater, Less};
     use std::vec::Vec;
